fix: arreglado el problema con un modulo que no encontraba pero sigo con problemas en la vista que tengo cosas que no van....
